# KAFKA-9685: Solve Set concatenation perf issue in AclAuthorizer