@jiao-zhangS 2020-03-10
> @jiao-zhangS Thanks for the PR. Can we update existing jmh benchmark test and benchmark authorize() method for some test input:
> https://github.com/apache/kafka/blob/trunk/jmh-benchmarks/src/main/java/org/apache/kafka/jmh/acl/AclAuthorizerBenchmark.java#L61

Sure, good point, let me update the benchmark test case.