@jiao-zhangS 2020-03-11
thanks, here is the perf result at local.
**before**
```
Benchmark                              (aclCount)  (resourceCount)  Mode  Cnt  Score   Error  Units
AclAuthorizerBenchmark.testAuthorizer           5             5000  avgt   15  0.167 ± 0.004  ms/op
AclAuthorizerBenchmark.testAuthorizer           5            10000  avgt   15  0.181 ± 0.002  ms/op
AclAuthorizerBenchmark.testAuthorizer           5            50000  avgt   15  0.157 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          10             5000  avgt   15  0.162 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          10            10000  avgt   15  0.149 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          10            50000  avgt   15  0.163 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          15             5000  avgt   15  0.149 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          15            10000  avgt   15  0.149 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          15            50000  avgt   15  0.166 ± 0.001  ms/op    
```
**after**
```
Benchmark                              (aclCount)  (resourceCount)  Mode  Cnt  Score    Error  Units
AclAuthorizerBenchmark.testAuthorizer           5             5000  avgt   15  0.037 ±  0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer           5            10000  avgt   15  0.042 ±  0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer           5            50000  avgt   15  0.034 ±  0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          10             5000  avgt   15  0.040 ±  0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          10            10000  avgt   15  0.037 ±  0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          10            50000  avgt   15  0.041 ±  0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          15             5000  avgt   15  0.038 ±  0.002  ms/op
AclAuthorizerBenchmark.testAuthorizer          15            10000  avgt   15  0.036 ±  0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          15            50000  avgt   15  0.036 ±  0.001  ms/op
```