@omkreddy 2020-03-12
There are checkStyle errors: `./gradlew :jmh-benchmarks:checkstyleMain`