@jiao-zhangS 2020-03-12
> There are checkStyle errors: `./gradlew :jmh-benchmarks:checkstyleMain`

oh.. Updated import control rule for benchmarks