@jiao-zhangS 2020-03-13
Thanks for review. One more thing,  as ratio of prefix acls to wildcard acls matters to the performance, I think it's better to cover more test cases. Modified the benchmark code a bit, WDYT? @omkreddy