@jiao-zhangS 2020-03-13
> makes sense. Can you share the latest benchmark results?
Sure, here are the results.

**before**
```
Benchmark                              (aclCount)  (resourceCount)  Mode  Cnt  Score   Error  Units
AclAuthorizerBenchmark.testAuthorizer           5             5000  avgt   15  0.223 ± 0.004  ms/op
AclAuthorizerBenchmark.testAuthorizer           5            10000  avgt   15  0.263 ± 0.009  ms/op
AclAuthorizerBenchmark.testAuthorizer           5            50000  avgt   15  1.249 ± 0.008  ms/op
AclAuthorizerBenchmark.testAuthorizer          10             5000  avgt   15  0.240 ± 0.010  ms/op
AclAuthorizerBenchmark.testAuthorizer          10            10000  avgt   15  0.245 ± 0.002  ms/op
AclAuthorizerBenchmark.testAuthorizer          10            50000  avgt   15  1.118 ± 0.046  ms/op
AclAuthorizerBenchmark.testAuthorizer          15             5000  avgt   15  0.217 ± 0.004  ms/op
AclAuthorizerBenchmark.testAuthorizer          15            10000  avgt   15  0.258 ± 0.011  ms/op
AclAuthorizerBenchmark.testAuthorizer          15            50000  avgt   15  1.132 ± 0.008  ms/op
```

**after**
```
Benchmark                              (aclCount)  (resourceCount)  Mode  Cnt  Score   Error  Units
AclAuthorizerBenchmark.testAuthorizer           5             5000  avgt   15  0.185 ± 0.002  ms/op
AclAuthorizerBenchmark.testAuthorizer           5            10000  avgt   15  0.235 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer           5            50000  avgt   15  0.420 ± 0.004  ms/op
AclAuthorizerBenchmark.testAuthorizer          10             5000  avgt   15  0.220 ± 0.009  ms/op
AclAuthorizerBenchmark.testAuthorizer          10            10000  avgt   15  0.222 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          10            50000  avgt   15  0.368 ± 0.013  ms/op
AclAuthorizerBenchmark.testAuthorizer          15             5000  avgt   15  0.193 ± 0.007  ms/op
AclAuthorizerBenchmark.testAuthorizer          15            10000  avgt   15  0.207 ± 0.002  ms/op
AclAuthorizerBenchmark.testAuthorizer          15            50000  avgt   15  0.422 ± 0.025  ms/op
```